Add basic code coverage using [tarpaulin][]
The coverage doesn't actually seem to work very well, but I think it's
better than nothing.  Tarpaulin reports the assertions in the solution
tests as themselves being uncovered. (Who watches the watchers?) It also
reports an `else` line in `bit_vec.rs` as uncovered -- not the code in
the actual then/else blocks, just the line with `} else {`.  It did
actually catch some uncovered code though, so I'm keeping it.

The bin/cov script is written in Bash, so not particularly portable.
Tarpaulin is currently only supported on Linux anyway, but it shouldn't
be hard to write a portable script once a portable coverage tool is
itself available.

[tarpaulin]: https://crates.io/crates/cargo-tarpaulin